Convert Backsnap tool to an operator. Fixes SKY-425, SKY-417.
The operator now runs continuously, reconciling PVCBackup objects by creating a
snapshot and backing it up.

Compared to the tool, the operator is more robust, allowing steps in the
process to fail and be retried later. Also, next to a default schedule for all
PVCs, it supports configurable backup schedules on a PVC and on a namespace,
allowing e.g. test namespaces to be backed up less often, or ephemeral PVCs not
to be backed up at all.

The tests are removed, since they don't apply anymore. They will be rewritten
and added back in a future commit.