Fix a case where backups stop because of a time difference.
In this case, a backup was scheduled for 12:00:00 UTC, and created just after
that time passed - but according to the apiserver's CreationTimestamp, it was
created at 11:59:59 UTC. The backup completed normally, but the automatic
operator wanted to create the next scheduled backup at 12:00:00 UTC again.
Because that scheduled backup had the same name as the latest one, further
creation failed.