Name backups after the current moment, not the scheduled moment.
Naming them after scheduled moment makes for nicer timestamps, and prevents
double creation of backup jobs if multiple controllers (or multiple workers)
are active. This is why I originally chose this approach. However, it also
causes issues when the job cannot be created because it already exists at the
scheduled time but something is wrong with it. Also, in the case of many missed
schedule moments, the name refers to the first missed one instead of the newly
made moment.

So, to make the controller more robust, I'm switching to naming them after the
current moment, not the scheduled moment.