Fix an issue where outdated cache would prevent a backup from being finished properly.
We'd mark a backup as Finished and immediately schedule another Reconcile, and
this second Reconcile would read the backup back as non-Finished due to caching
in the controller runtime. This would cause the backup to be considered running
again, but the second Status update would fail because the object is outdated.
This second Reconcile, then, would never be rescheduled, even though it should
be. To prevent this altogether, make sure that we wait for the object to be
read back properly, as suggested by the controller runtime developers.

https://github.com/kubernetes-sigs/controller-runtime/issues/1464