Set created Pods to never retry automatically.
When a Pod fails, it is currently set to retry automatically. The Job
controller currently allows 4 retries (the BackoffLimit). After 4 retries in
the Pod, the Job controller must prevent further retries and therefore deletes
the Pod. Because of this, we then also lose the Pod logs, so if they weren't
caught by monitoring, we lose the reason the Job failed.

With this change, the Pod itself only retries once. The Job controller supports
this, and instead, creates new Pods for further retries, but leaves the old
Pods around. With this mechanism, we get the same amount of retries, but old
logs can still be inspected while the Job exists.